+++ CHANGELOG.md
@@ -14,10 +14,12 @@ This CHANGELOG follows the format listed at [Keep A Changelog](http://keepachang
 - syntax error in check-sensu-clients
 
 ### Added
+- check-rds.rb: Add support for assuming a role in another account
 - check-instance-events.rb: Add instance_id option
 - check-sensu-clients.rb: SSL support
 - common.rb: adding support for environment variable AWS_REGION when region is specified as an empty string
 - metrics-sqs.rb: Add support for recording additional per-queue SQS metrics (counts of not-visible and delayed messages)
+- check-subnet-ip-consumption.rb: to check consumption of IP addresses in subnets and alert if consumption exceeds a threshold
 - check-beanstalk-health.rb: Add optional region support
 
 ## [3.1.0] - 2016-05-15